Add some contrast to subreddit and user toolbars.
Add a gradient from the primary theme color to transparent so that if a
subreddit or user profile has a very light background, the usually light
text and buttons are not obscured or in some cases invisible.

The gradients don't appear for immersive mode so that the app remains
immersive.