Store only first level replies as children (#1222)
* Show the current image index in gallery in PostFragment.

* Fixed issue where filters applied to same feed were behaving incorrectly (#1172)

* Fix ItemTouchHelper and gallery swipe gesture fighting each other.

* Show correct image in the gallery in ViewRedditGalleryActivity.

* Swipe to view images in a gallery in Card Layout 2.

* Fix gallery layout issues.

* Remove `commentData` argument that is always an empty ArrayList

* Return top level comments in addition to expanded comments after parsing

Since 0f1c4d loading more comments loads not only first level replies, but also deeper comments. Because of this `expandedComments` can contain those deep replies if `expandChildren` is true. Adding `expandedComments` to parent causes a bug because parent's children are supposed to be only next level replies. Because of previously mentioned changes that is not true.

Now expanding parent comment results in duplicate comments: one of them correctly comes from the parent of duplicated comment. The other one is shown because it is incorrectly stored in the parent of "load more comments" button.

This comment separates top level comments (fist level replies) and expanded comments. `expandedComments` are still used for display, but only first level replies are added to the parent

* Add debug assertion for children depth

Co-authored-by: Docile-Alligator <25734209+Docile-Alligator@users.noreply.github.com>
Co-authored-by: Aidan223 <110802888+Aidan223@users.noreply.github.com>